move some IC traits to client
